+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes to this package will be documented in this file. The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
+## [1.8.6] - 2023-10-05
+
+### Changed
+
+- In the `Create Asset` menu, the `Signal` and `Timeline` items have been moved to the `Timeline` submenu.
+
+### Fixed
+
+- Fixed issue where changing the name of a group track was not undoable ([TB-218](https://issuetracker.unity3d.com/issues/timeline-track-group-renaming-doesnt-go-into-the-undo-history)).
+- Fixed issue where a warning would be logged in the console for `TrackAsset` ([TB-229](https://issuetracker.unity3d.com/product/unity/issues/guid/TB-229)).
+- Fixed performance regression when rebuilding the playable graph.
+- Fixed a `MissingReferenceException` when editing a Timeline that is an instance of a prefab and a bound `Animator` is destroyed ([TB-222](https://issuetracker.unity3d.com/issues/timeline-missingreferenceexceptions-when-undoing-add-animator-from-timeline))
+- Removed warnings due to obsolete analytics API.
+
 ## [1.8.5] - 2023-08-24
 
 ### Added